Extract some methods from phased_fsim.py to workflow.py
