+++ cirq/google/calibration/__init__.py
@@ -4,7 +4,9 @@ from cirq.google.calibration.phased_fsim import (
     FloquetPhasedFSimCalibrationResult,
     PhasedFSimParameters,
     PhasedFSimCalibrationRequest,
-    PhasedFSimCalibrationResult,
+    PhasedFSimCalibrationResult
+)
+from cirq.google.calibration.workflow import (
     floquet_calibration_for_circuit,
     floquet_calibration_for_moment,
     run_calibrations,
